fix(xilinx): resolve misra rule 12.2 violations
Fixed below MISRA violation:
- MISRA Violation: MISRA-C:2012 R.12.2:
  - The right hand operand of a shift operator shall lie in the range
    zero to one less than the width in bits of the essential type of
    the left hand operand.
- Fix:
  - Type casted left operand to a larger width than shift.

+++ plat/xilinx/common/include/pm_common.h
@@ -31,7 +31,7 @@
 
 #define TZ_VERSION_MAJOR	1U
 #define TZ_VERSION_MINOR	0U
-#define TZ_VERSION		((TZ_VERSION_MAJOR << 16U) | \
+#define TZ_VERSION		(((uint32_t)TZ_VERSION_MAJOR << 16U) | \
 				 TZ_VERSION_MINOR)
 
 /**
